hw4 fixes
