+++ README.md
@@ -1,0 +1,6 @@
+Старт сервер сокета с помощью команды:
+php socet_server_start.php > from_client_message.txt
+
+Сообщения которые приходят от клиента пишутся в файл from_client_message.txt
+
+Запрсы от клиента далаются командой php socet_client.php , сервер отвечает в консоль
