+++ socet_server_start.php
@@ -1,0 +1,8 @@
+<?php
+require('./classes/SocketServer.php');
+$socketServer = new SocketServer();
+try {
+    $socketServer->run();
+} catch (Exception $e) {
+    echo 'Выброшено исключение: ', $e->getMessage(), "\n";
+}
